add: polygon auth

+++ src/main/kotlin/cn/xlor/xloj/model/Users.kt
@@ -5,10 +5,11 @@ import org.ktorm.entity.Entity
 import org.ktorm.entity.sequenceOf
 import org.ktorm.schema.Table
 import org.ktorm.schema.int
+import org.ktorm.schema.long
 import org.ktorm.schema.text
 
 data class UserProfile(
-  val id: Int,
+  val id: Long,
   val username: String,
   val nickname: String
 )
@@ -16,7 +17,7 @@ data class UserProfile(
 interface User : Entity<User> {
   companion object : Entity.Factory<User>()
 
-  var id: Int
+  var id: Long
 
   var username: String
 
@@ -28,7 +29,7 @@ interface User : Entity<User> {
 fun User.toUserProfile() = UserProfile(id, username, nickname)
 
 object Users : Table<User>("users") {
-  val id = int("id").primaryKey().bindTo { it.id }
+  val id = long("id").primaryKey().bindTo { it.id }
 
   val username = text("username").bindTo { it.username }
